op-devstack: fix L2CL from sequencer to verifier
