+++ op-acceptance-tests/tests/sync_tester_ext_el/sync_tester_ext_el_test.go
@@ -4,7 +4,6 @@ import (
 	"testing"
 
 	"github.com/ethereum-optimism/optimism/op-devstack/devtest"
-	"github.com/ethereum-optimism/optimism/op-devstack/dsl"
 	"github.com/ethereum-optimism/optimism/op-devstack/presets"
 	"github.com/ethereum-optimism/optimism/op-service/eth"
 	"github.com/ethereum-optimism/optimism/op-supervisor/supervisor/types"
@@ -25,25 +24,13 @@ func TestSyncTesterExtEL(gt *testing.T) {
 	require.NotNil(sys.L2CL, "L2 CL node should be available")
 	require.NotNil(sys.SyncTester, "SyncTester should be available")
 
-	// Test that we can get chain ID from SyncTester
-	syncTester := sys.SyncTester.Escape()
-	syncTesterChainID, err := syncTester.API().ChainID(t.Ctx())
-	require.NoError(err, "should be able to get chain ID from SyncTester")
-	require.Equal(eth.ChainIDFromUInt64(11155420), syncTesterChainID, "SyncTester should be on chain 11155420")
-
-	// Test that L2CL node and SyncTester are on the same chain
-	require.Equal(l2CLChainID, syncTesterChainID, "L2CL node and SyncTester should be on the same chain")
-
 	// Test that we can get sync status from L2CL node
 	l2CLSyncStatus := sys.L2CL.SyncStatus()
 	require.NotNil(l2CLSyncStatus, "L2CL should have sync status")
 
-	dsl.CheckAll(t,
-		sys.L2CL.AdvancedFn(types.LocalUnsafe, 31987002, 300),
-	)
+	sys.L2CL.Advanced(types.LocalUnsafe, 32012768, 1000)
 
 	t.Logger().Info("SyncTester ExtEL test completed successfully",
 		"l2cl_chain_id", l2CLChainID,
-		"sync_tester_chain_id", syncTesterChainID,
 		"l2cl_sync_status", l2CLSyncStatus)
 }
