+++ op-devstack/sysgo/l2_network_superchain_registry.go
@@ -3,7 +3,6 @@ package sysgo
 import (
 	"fmt"
 
-	"github.com/davecgh/go-spew/spew"
 	"github.com/ethereum/go-ethereum/common"
 	"github.com/ethereum/go-ethereum/core"
 
@@ -36,14 +35,6 @@ func WithL2NetworkFromSuperchainRegistry(l2NetworkID stack.L2NetworkID, networkN
 		genesis := &core.Genesis{
 			Config: paramsChainConfig,
 		}
-
-		fmt.Println("==== anteva genesis for L2 - start")
-		spew.Dump(genesis)
-		fmt.Println("==== anteva genesis for L2 - end")
-
-		fmt.Println("==== anteva rollupcfg for L2 - start")
-		spew.Dump(rollupCfg)
-		fmt.Println("==== anteva rollupcfg for L2 - end")
 
 		// Create a deployment config (this will be minimal since we're using registry configs)
 		deployment := &L2Deployment{
